requirements.txt
async-timeout==4.0.2
blinker==1.6.2
click==8.1.6
Deprecated==1.2.14
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
packaging==23.1
platformdirs==3.9.1
pluggy==1.2.0
pyaml==23.7.0
pytest==7.4.0
PyYAML==6.0.1
redis==4.2.2
virtualenv==20.24.2
Werkzeug==2.3.6
wrapt==1.15.0
Yapsy==1.12.2
